+++ app/src/androidTest/java/woowacourse/movie/presentation/utils/ClickUtils.kt
@@ -1,0 +1,10 @@
+package woowacourse.movie.presentation.utils
+
+import androidx.test.espresso.ViewInteraction
+import androidx.test.espresso.action.ViewActions
+
+fun ViewInteraction.repeatClick(count: Int) {
+    repeat(count) {
+        perform(ViewActions.click())
+    }
+}
