+++ app/src/test/java/woowacourse/movie/presentation/ui/reservation/ReservationPresenterTest.kt
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import woowacourse.movie.domain.model.Reservation
 import woowacourse.movie.presentation.fakerepository.FakeReservationRepository
-import woowacourse.movie.presentation.ui.getDummyReservation
 import woowacourse.movie.presentation.ui.reservation.fake.FakeReservationView
+import woowacourse.movie.presentation.utils.getDummyReservation
 
 class ReservationPresenterTest {
     private lateinit var reservation: Reservation
